Add unit test for matching grounded atoms
Move dictionary matching test into examples.
